+++ packages/desktop/src/components/home/TransactionTable.svelte
@@ -3,7 +3,12 @@
     import { transactionStore } from "$lib/transaction.svelte";
     import Transaction from "./transaction/Transaction.svelte";
 
-    const tableStore = new TableStore();
+    interface Props {
+        tableStore: TableStore
+    }
+
+    const { tableStore }: Props = $props();
+    let selected = $derived(tableStore.allRowsSelected);
 </script>
 
 <table>
@@ -20,7 +25,7 @@
     <thead>
         <tr>
             <th>
-                <input type="checkbox" name="selected" id="row-checkbox"
+                <input type="checkbox" checked={selected} name="selected" id="row-checkbox"
                 onclick={(e) => {
                     if(e.currentTarget.checked){
                         tableStore.toggleSelectAll();